+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * First release.
 
 ## 1.0.2
-* SnackbarType added on showSnackbar
+* Type added on showSnackbar
 
 ## 1.0.3
 * Snackbar customization properties added
@@ -12,3 +12,18 @@
 
 ## 1.0.5
 * Code formating and pub score related changes
+
+## 1.0.6
+
+### Improvements & Fixes
+- Naming convention fixes.
+
+### Breaking Changes
+- Renamed `SnackbarType` to `Type`.
+- Renamed `ToastType` to `Type`.
+- Renamed `LoadingIndicatorPosition` to `Position`.
+
+### Default Values Updated
+- **Snackbar type:** `Info`
+- **Snackbar position:** `Top`
+- **Toast position:** `Top`